Updated Julia solid-body code
